// frontend/client/model/authResponses.ts
/**
 * angular-typescript-starter
 * BE
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AuthPayload } from './authPayload';


export interface AuthResponses { 
    success: boolean;
    payload?: AuthPayload;
    message?: any | null;
}