// frontend/client/model/userPayload.ts
/**
 * angular-typescript-starter
 * BE
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PickIUserFirstnameOrLastnameOrUsernameOrEmailOrUserTypeOrCreatedAtOrUpdatedAt } from './pickIUserFirstnameOrLastnameOrUsernameOrEmailOrUserTypeOrCreatedAtOrUpdatedAt';


export interface UserPayload { 
    user?: PickIUserFirstnameOrLastnameOrUsernameOrEmailOrUserTypeOrCreatedAtOrUpdatedAt;
    users?: Array<PickIUserFirstnameOrLastnameOrUsernameOrEmailOrUserTypeOrCreatedAtOrUpdatedAt>;
    userLength?: number;
    token?: string;
}